Add CQRS to Product service

+++ product-service/src/main/java/fr/myitworld/productservice/entities/Product.java
@@ -9,6 +9,7 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import java.math.BigDecimal;
 
 @Entity
 @Data
@@ -21,7 +22,7 @@ public class Product {
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
     private String name;
-    private Double price;
-    private Long quantity;
+    private BigDecimal price;
+    private Integer quantity;
 
 }
